Split src into main and tests
