+++ src/main/utils/SettingsUtility.js
